+++ example/api/global-interfaces/includes/builtin/test.sh
@@ -1,0 +1,4 @@
+#!/bin/bash
+
+xmake config --verbose --clean
+xmake build --verbose