+++ xmake/actions/config/xmake.lua
@@ -226,7 +226,6 @@ task("config")
                     {nil, "debugger",   "kv", "auto"    , "Set debugger"},
                     {nil, "ccache",     "kv", true      , "Enable or disable the c/c++ compiler ccache."},
                     {nil, "xcache",     "kv", true      , "Enable or disable the c/c++ compiler xmake cache."},
-                    {nil, "debug",      "kv", false     , "Enable or disable to print debug information."},
                     {nil, "xcachedir",  "kv", nil       , "Set the xcache directory."},
                     {nil, "trybuild",   "kv", nil       , "Enable try-build mode and set the third-party buildsystem tool.",
                                                             "e.g.",